♻️  (HomeButton) refactor home button
Create separate component for home button to reuse it in other interfaces
