+++ box-ui/src/components/css/HomeButtonsComponent.css
@@ -2,15 +2,6 @@
     padding-top: 4%;
 }
 
-.HomeButton {
-    height: 17vh;
-    width: 60vw;
-    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-    overflow-wrap: 'break-word';
-    overflow-x: 'visible';
-    overflow-y: 'hidden';
-}
-
 #counter {
     color: azure;
     margin-top: 1%;
